feat(pci-volume-snapshot): add missing config files
ref: DTCORE-3209
