+++ package.json
@@ -95,6 +95,7 @@
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-markdown": "^1.0.0",
     "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "esm": "^3.2.25",
     "execa": "^2.0.5",
