+++ packages/manager/apps/pci-volume-snapshot/.eslintrc.cjs
@@ -1,0 +1,27 @@
+module.exports = {
+  extends: [
+    '../../../../.eslintrc.js',
+    'plugin:react/recommended',
+    'plugin:react/jsx-runtime',
+    'plugin:react-hooks/recommended-legacy',
+  ],
+  settings: {
+    react: {
+      version: 'detect',
+    },
+  },
+  globals: {
+    __VERSION__: true,
+  },
+  rules: {
+    'import/extensions': 'off',
+    '@typescript-eslint/no-non-null-assertion': 'off',
+    'arrow-body-style': ['error', 'as-needed'],
+    'prefer-const': 'error',
+    'react/jsx-curly-brace-presence': [
+      'error',
+      { props: 'never', children: 'never' },
+    ],
+    'react/jsx-boolean-value': ['error', 'never'],
+  },
+};
