+++ packages/manager/apps/pci-volume-snapshot/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@ovh-ux/manager-config": "^8.0.2",
     "@ovh-ux/manager-core-api": "^0.10.1",
-    "@ovh-ux/manager-core-utils": "*",
+    "@ovh-ux/manager-core-utils": "^0.3.0",
     "@ovh-ux/manager-pci-common": "^2.3.1",
     "@ovh-ux/manager-react-components": "^2.15.1",
     "@ovh-ux/manager-react-core-application": "^0.11.13",
@@ -55,6 +55,7 @@
     "@vitest/coverage-v8": "^1.6.0",
     "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "postcss": "^8.4.38",
     "rollup": "^4.18.0",
     "tailwindcss": "^3.4.4",
