+++ packages/manager/apps/pci-volume-snapshot/src/core/HidePreloader.tsx
@@ -6,6 +6,7 @@ export default function HidePreloader() {
 
   useEffect(() => {
     ux.hidePreloader();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return null;
