+++ packages/manager/apps/pci-volume-snapshot/src/pages/listing/useDatagridColumn.tsx
@@ -5,10 +5,16 @@ import {
 } from '@ovh-ux/manager-react-components';
 import { useTranslation } from 'react-i18next';
 import { useBytes } from '@ovh-ux/manager-pci-common';
-import { TVolumeSnapshot } from '@/api/hooks/useSnapshots';
+import { TVolumeSnapshot } from '@/api/api.types';
 import { useFormattedDate } from '@/hooks/useFormattedDate';
 import StatusComponent from './Status.component';
 import ActionsComponent from './Actions.component';
+
+const CreationDateCell = (props: TVolumeSnapshot) => (
+  <DataGridTextCell>
+    {useFormattedDate(props.creationDate, 'P p')}
+  </DataGridTextCell>
+);
 
 export const useDatagridColumn = () => {
   const { t } = useTranslation('volumes');
@@ -61,11 +67,7 @@ export const useDatagridColumn = () => {
     {
       id: 'creationDate',
       label: t('pci_projects_project_storages_snapshots_creationDate_label'),
-      cell: (props: TVolumeSnapshot) => (
-        <DataGridTextCell>
-          {useFormattedDate(props.creationDate, 'P p')}
-        </DataGridTextCell>
-      ),
+      cell: CreationDateCell,
     },
     {
       id: 'status',
