+++ packages/manager/apps/pci-volume-snapshot/src/setupTests.tsx
@@ -1,6 +1,8 @@
 import '@testing-library/jest-dom';
 import 'element-internals-polyfill';
+import { OdsInputChangeEvent } from '@ovhcloud/ods-components';
 import { vi } from 'vitest';
+import React from 'react';
 
 vi.mock('react-router-dom', async () => {
   const mod = await vi.importActual('react-router-dom');
@@ -8,7 +10,7 @@ vi.mock('react-router-dom', async () => {
     ...mod,
     useSearchParams: () => [new URLSearchParams({})],
     useParams: () => ({ projectId: 'project-id', kubeId: 'kube-id' }),
-    useHref: vi.fn(),
+    useHref: vi.fn().mockImplementation((to: string) => to),
     useLocation: vi.fn(),
     useNavigate: vi.fn(),
     Navigate: () => null,
@@ -23,7 +25,11 @@ vi.mock('@ovh-ux/manager-pci-common', async () => {
     useProject: vi.fn().mockResolvedValue({
       projectName: 'project-name',
       project_id: 'project-id',
+      data: { description: 'Test Project' },
     }),
+    PciAnnouncementBanner: () => (
+      <div data-testid="announcement-banner">Announcement Banner</div>
+    ),
   };
 });
 
@@ -49,15 +55,164 @@ vi.mock('@ovh-ux/manager-react-components', async () => {
   };
 });
 
+vi.mock('@ovhcloud/ods-components/react', async (importOriginal) => {
+  const actual = (await importOriginal()) || {};
+  return {
+    ...(typeof actual === 'object' && actual !== null ? actual : {}),
+    OdsButton: ({
+      label,
+      icon,
+      isDisabled,
+      onClick,
+      className,
+      'data-testid': dataTestId,
+    }: {
+      label: string;
+      icon: string;
+      isDisabled: string | boolean;
+      onClick: () => void;
+      className: string;
+      'data-testid': string;
+    }) => (
+      <button
+        type="button"
+        className={className}
+        disabled={isDisabled === 'true' || isDisabled === true}
+        onClick={onClick}
+        data-testid={dataTestId}
+      >
+        {icon && <span className={`icon-${icon}`} />}
+        {label}
+      </button>
+    ),
+    OdsInput: ({
+      name,
+      value,
+      onOdsChange,
+      onKeyDown,
+      className,
+      'data-testid': dataTestId,
+    }: {
+      name: string;
+      value: string;
+      onOdsChange: (e: Partial<OdsInputChangeEvent>) => void;
+      onKeyDown: () => void;
+      className: string;
+      'data-testid': string;
+    }) => (
+      <input
+        type="text"
+        name={name}
+        className={className}
+        value={value}
+        onChange={(e) => {
+          onOdsChange?.({
+            detail: {
+              value: e.target.value,
+              name,
+            },
+            stopPropagation: () => {},
+            preventDefault: () => {},
+          });
+        }}
+        onKeyDown={onKeyDown}
+        data-testid={dataTestId}
+      />
+    ),
+    OdsSelect: ({
+      name,
+      value,
+      onOdsChange,
+      children,
+      'data-testid': dataTestId,
+    }: {
+      name: string;
+      value: string;
+      children: React.ReactNode;
+      onOdsChange: (e: Partial<OdsInputChangeEvent>) => void;
+      'data-testid': string;
+    }) => (
+      <select
+        name={name}
+        value={value}
+        onChange={(e) => {
+          onOdsChange?.({
+            detail: { value: e.target.value, name },
+            stopPropagation: () => {},
+            preventDefault: () => {},
+          });
+        }}
+        data-testid={dataTestId}
+      >
+        {children}
+      </select>
+    ),
+    OdsBadge: ({
+      color,
+      label,
+      'data-testid': testId,
+    }: {
+      color: string;
+      label: string;
+      'data-testid': string;
+    }) => (
+      <div data-testid={testId} data-color={color}>
+        {label}
+      </div>
+    ),
+  };
+});
+
 vi.mock('react-i18next', () => ({
   useTranslation: () => ({
     t: (translationKey: string) => translationKey,
     i18n: {
       changeLanguage: () => new Promise(() => {}),
       exists: () => true,
+      language: 'en-US',
     },
   }),
   Trans: ({ children }: { children: string }) => children,
+}));
+
+vi.mock('date-fns', async (importOriginal) => {
+  const actual = (await importOriginal()) as Record<string, unknown>;
+  return {
+    ...actual,
+    format: vi.fn((date, formatStr, options) => {
+      // Simple mock implementation
+      const locale = options?.locale?.code || 'en';
+      const dateObj = new Date(date);
+
+      // Return a predictable format for testing
+      if (formatStr === 'PPpp') {
+        return `${dateObj.toLocaleDateString()}, ${dateObj.toLocaleTimeString()} (${locale})`;
+      }
+      if (formatStr === 'P p') {
+        return `${dateObj.toLocaleDateString()} ${dateObj.toLocaleTimeString()} (${locale})`;
+      }
+      return `${dateObj.toISOString()} (${formatStr}) (${locale})`;
+    }),
+  };
+});
+
+vi.mock('date-fns/locale', () => ({
+  enUS: { code: 'en' },
+  fr: { code: 'fr' },
+  de: { code: 'de' },
+  es: { code: 'es' },
+}));
+
+vi.mock('@ovh-ux/manager-core-utils', () => ({
+  getDateFnsLocale: (language: string) => {
+    const localeMap: Record<string, string> = {
+      'en-US': 'enUS',
+      'fr-FR': 'fr',
+      'de-DE': 'de',
+      'es-ES': 'es',
+    };
+    return localeMap[language] || 'enUS';
+  },
 }));
 
 vi.mock('@ovh-ux/manager-react-shell-client', async () => {
@@ -68,6 +223,13 @@ vi.mock('@ovh-ux/manager-react-shell-client', async () => {
       trackPage: vi.fn(),
       trackClick: vi.fn(),
     })),
+    ShellContext: React.createContext({
+      environment: {
+        getUser: () => ({
+          ovhSubsidiary: 'FR',
+        }),
+      },
+    }),
   };
 });
