+++ packages/manager/apps/pci-volume-snapshot/src/wrapperRenders.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import {
   ShellContext,
